+++ src/services/recruit/apis/index.ts
@@ -9,6 +9,7 @@ export * from './getRecruitApplication';
 export * from './getRecruitDetail';
 export * from './getRecruitParticipants';
 export * from './getRecruits';
+export * from './likeRecruitApplication';
 export * from './removeRecruit';
 export * from './rejectRecruitApplication';
 export * from './scrapRecruit';
