+++ src/services/recruit/hooks/types.ts
@@ -1,0 +1,4 @@
+export interface RecruitApplicationParams {
+  recruitId: number;
+  recruitApplicationId: number;
+}
