+++ src/services/recruit/hooks/useCompleteRecruit.ts
@@ -1,0 +1,9 @@
+import { useMutation } from '@tanstack/react-query';
+
+import { completeRecruit } from '~/services/recruit/apis';
+
+export const useCompleteRecruit = (recruitId: number) => {
+  return useMutation({
+    mutationFn: () => completeRecruit(recruitId),
+  });
+};
