+++ src/services/recruit/hooks/useCreateRecruit.ts
@@ -1,0 +1,9 @@
+import { useMutation } from '@tanstack/react-query';
+
+import { createRecruit } from '~/services/recruit/apis';
+
+export const useCreateRecruit = () => {
+  return useMutation({
+    mutationFn: createRecruit,
+  });
+};
