+++ src/services/recruit/hooks/useRemoveRecruit.ts
@@ -1,0 +1,9 @@
+import { useMutation } from '@tanstack/react-query';
+
+import { removeRecruit } from '~/services/recruit/apis';
+
+export const useRemoveRecruit = (recruitId: number) => {
+  return useMutation({
+    mutationFn: () => removeRecruit(recruitId),
+  });
+};
